cleanup theme files
