+++ src/client/App.tsx
@@ -3,10 +3,11 @@ import { createRoot } from 'react-dom/client';
 import localStorageAvailable from '../util/localStorageAvail';
 import Board from './components/game_elements/board';
 import CssBaseline from '@mui/material/CssBaseline';
-import { ThemeProvider, createTheme } from '@mui/material/styles';
+import { ThemeProvider } from '@mui/material/styles';
+import { lightTheme } from './Theme';
+import { darkTheme } from './Theme';
 import { Box, useTheme } from '@mui/material/';
 import Header from './components/Header';
-
 import useMediaQuery from '@mui/material/useMediaQuery';
 import Leaderboard from './components/leaderboard';
 import { AuthProvider } from './context/AuthContext';
@@ -31,31 +32,6 @@ function initialState() {
 function App() {
   const [darkMode, setDarkMode] = useState<boolean>(initialState());
 
-  const lightTheme = createTheme({
-    palette: {
-      mode: 'light',
-      primary: {
-        main: '#E11774',
-        dark: '#46A4DB',
-        contrastText: '#fff',
-      },
-    },
-    typography: {
-      fontFamily: ['Roboto', 'Lalezar', 'Poppins'].join(','),
-    },
-  });
-
-  const darkTheme = createTheme({
-    palette: {
-      mode: 'dark',
-      primary: {
-        main: '#05FFF4',
-        dark: '#46A4DB',
-        contrastText: '#000',
-      },
-    },
-  });
-
   const theme = useTheme();
 
   const isMobile = useMediaQuery(theme.breakpoints.down('sm'));
@@ -75,7 +51,7 @@ function App() {
           width="100vw"
           gap={2}
           justifyContent={'center'}
-          flexDirection={isMobile ? 'column' : 'row'}
+          flexDirection={isMobile ? 'column' : 'row'} // start with this
         >
           <Board darkMode={darkMode} />
           <Leaderboard />
