+++ src/client/Theme/index.ts
@@ -1,0 +1,2 @@
+export { default as lightTheme } from './LightTheme';
+export { default as darkTheme } from './DarkTheme';
